requirements.txt

streamlit
tensorflow-cpu
pillow
requests
numpy